iommu/riscv: Prepare to use MSI table
Capture the IMSIC layout from its config and reserve all the addresses.
Then use the IMSIC layout info to calculate the maximum number of PTEs
the MSI table needs to support and allocate the MSI table when creating
the IRQ domain. Finally, map the IMSIC addresses in the stage1 DMA
table when the stage1 DMA table is not BARE. This ensures it doesn't
fault as it will translate the addresses before the MSI table does.

Signed-off-by: Andrew Jones <ajones@ventanamicro.com>